requirements.txt
alembic==1.7.5
argcomplete==2.0.0
asgiref==3.5.2
Authlib==1.0.1
Babel==2.9.1
bcrypt==3.2.2
Beaker==1.10.0
beautifulsoup4==4.11.0
bidict==0.21.4
blinker==1.4
blivet==3.4.4
blivet-gui==2.3.0
boto3==1.21.33
botocore==1.24.33
Brlapi==0.8.4
Brotli==1.0.9
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
chrome-gnome-shell==0.0.0
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
cupshelpers==1.0
dasbus==1.6
dbus-python==1.2.18
decorator==5.1.1
distro==1.6.0
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
dnspython==2.2.0
docopt==0.6.2
drf-yasg==1.20.0
email-validator==1.1.1
eventlet==0.33.0
factory-boy==3.2.1
Faker==13.11.1
fedora-third-party==0.10
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.2
Flask-SocketIO==5.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fros==1.1
gpg==1.15.1
greenlet==1.1.2
gssapi==1.7.2
httpagentparser==1.9.2
humanize==3.13.1
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
koji==1.29.1
langtable==0.0.58
ldap3==2.8.1
libcomps==0.1.18
lxml==4.7.1
Mako==1.1.4
MarkupSafe==2.0.0
mod-wsgi==4.9.0
msgpack==1.0.3
nftables==0.1
olefile==0.46
packaging==21.3
paramiko==2.11.0
passlib==1.7.4
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==9.1.0
ply==3.11
productmd==1.33
progressbar2==3.53.2
prompt-toolkit==3.0.29
psutil==5.8.0
psycopg2==2.9.3
ptyprocess==0.6.0
pwquality==1.4.4
pyasn1==0.4.8
pycairo==1.21.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.1
pyenchant==3.2.2
pyflakes==2.4.0
pygit2==1.7.1
Pygments==2.12.0
PyGObject==3.42.1
pykickstart==3.36
PyNaCl==1.4.0
pynvim==0.4.3
pyOpenSSL==21.0.0
pyotp==2.6.0
pyparsing==2.4.7
pyparted==3.12.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.1
python-engineio==4.3.2
python-meh==0.50
python-socketio==5.6.0
python-utils==2.5.6
pytz==2022.1
pyudev==0.22.0
pyvim==3.0.2
pyxdg==0.27
qrcode==7.3.1
regex==2022.6.2
requests==2.27.1
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
rpm==4.17.0
rpmautospec==0.2.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.1
selinux==3.3
sepolicy==3.3
setools==4.4.0
simpleaudio==1.0.4
simplejson==3.17.6
simpleline==1.9.0
six==1.16.0
sos==4.3
soupsieve==2.3.1
SQLAlchemy==1.4.39
sqlparse==0.4.2
sshtunnel==0.4.0
systemd-python==234
Tempita==0.5.2
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.8
user-agents==2.2.0
wcwidth==0.2.5
Werkzeug==2.0.3
WTForms==3.0.1
